InstrumentStatus component
