+++ src/Components/InstrumentList.tsx
@@ -3,24 +3,14 @@ import {Link, useLocation} from "react-router-dom";
 import {Instrument} from "../../Interfaces";
 import dateFormatter from "dayjs";
 import {filter} from "lodash";
+import InstrumentStatus from "./InstrumentStatus";
 
 interface Props {
     instrumentList: Instrument[],
     listError: string
 }
 
-function getStatusColor(status: string | undefined) {
-    switch (status) {
-        case "Active":
-            return "success";
-        case "Erroneous":
-            return "error";
-        case "Failed":
-            return "error";
-        default:
-            return "info";
-    }
-}
+
 
 function InstrumentList({instrumentList, listError}: Props): ReactElement {
 
@@ -117,10 +107,7 @@ function InstrumentList({instrumentList, listError}: Props): ReactElement {
                                         {item.fieldPeriod}
                                     </td>
                                     <td className="table__cell ">
-                                        <span
-                                            className={`status status--${getStatusColor(item.status)}`}>
-                                            {item.status}
-                                        </span>
+                                        <InstrumentStatus status={item.status ? item.status: ""}/>
                                     </td>
                                     <td className="table__cell ">
                                         {dateFormatter(item.installDate).format("DD/MM/YYYY HH:mm")}
